pyproject.toml
[project]
name = "jwt-celery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.1.2",
    "celery>=5.5.3",
    "fastapi>=0.116.1",
    "google-generativeai>=0.5.4",
    "langchain==0.3.27",
    "langchain-community>=0.2.5",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=1.0.6",
    "langchain-openai>=0.1.14",
    "langchain-qdrant>=0.2.0",
    "motor==3.3.2",
    "openai>=1.101.0",
    "passlib[bcrypt]==1.7.4",
    "pydantic==2.11.7",
    "pydantic-settings>=2.2.1",
    "pymongo==4.6.0",
    "pypdf2>=3.0.1",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-magic>=0.4.27",
    "python-multipart==0.0.6",
    "qdrant-client>=1.12.1",
    "redis>=6.4.0",
    "tiktoken>=0.11.0",
    "uvicorn==0.24.0",
]
